Moved audit logic to common package, generalised audit interaction id.
