--- src/main/java/uk/nhs/cdss/audit/sqs/AuditSender.java
@@ -1,7 +1,0 @@
-package uk.nhs.cdss.audit.sqs;
-
-import uk.nhs.cdss.audit.model.AuditSession;
-
-public interface AuditSender {
-    void sendAudit(AuditSession session);
-}
