+++ src/main/java/uk/nhs/cdss/config/AuditConfig.java
@@ -1,0 +1,10 @@
+package uk.nhs.cdss.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan("uk.nhs.cactus.common.audit")
+public class AuditConfig {
+
+}
